requirements.txt
influxdb-client==1.24.0
jmespath==0.9.4
python-dateutil==2.8.1
requests==2.22.0
urllib3==1.25.8
pandas==1.3.5
octopus-energy-client
